fixes
careful of patch fix on cost having indexing of int

+++ mav_state.py
@@ -114,7 +114,7 @@ class MAV_State:
                                [float(self.chi)],
                                [float(self.p)],
                                [float(self.q)],
-                               [float(self.r)]])
+                               [float(self.r)]], dtype=np.float32)
 
         return this_state
